Use Jinja2 templates instead of string formatting.
I intentionally chose to use string formatting initially because I was trying to keep things simple. I wanted the server to be a single file for easier deployment and to have fewer dependencies. But in hindsight I think it's not worth the tradeoffs. Using templates makes development easier and reduces opportunity for errors.

I picked Jinja2 semi arbitrarily. I have experience with it and like it, and it showed up near the top of a Google search. I didn't investigate whether another template library would be better.