Fix the "Check" button never being enabled.
This fixes issue #4. Thanks to citronalco for bringing it up!

I had fixed this years ago on my server where I was running it but apparently I neglected to check in the fix. I recently moved to a new server and lost the fix, so it was broken on kingant.net.